bookr-mod 06-12
2009-6-12

Display PDF chapters directory, submitted by cg friends xyhchina. After
selecting the directory, press O to jump to the specified position and
adjust the page scale according to the zoom information in the
directory; press the key to jump to the specified position and maintain
the original page scale, or maintain the original zoom ratio and
Horizontal position (in the Options button can be set whether to ignore
the horizontal position information, you can also in the Outlines
interface, press the Trigger key to temporarily switch).
Fixed a question of Japanese space. The previous version of the compiler
forgot to update the modified cmap data, resulting in the problem still
exists.
Fixed an error in the text of PDF.
Add single and double page mode. Valid only for PDF and DJVU. You can
use the top button in the fourth column of the Select menu to toggle.
When the mode is on, the page will automatically remember the horizontal
position according to the number of parity pages.
In the Select menu, increase the display of the current file title (when
the title does not show the file name) and single and double page mode
status.
Optimize the rendering speed of the file selection dialog.
Adjust the key response speed in the file selection dialog box.
Optimize the font cache to raise the loading speed for pages that
contain large numbers of Chinese characters.
Fixed long file name display incomplete problem, currently supports the
display of the maximum length of the memory stick file name.
Fix some of the characters in the file name to block the previous
character.
Optimize page loading, scaling, and rotation speed when the CPU
frequency is set to a lower value.
Options Adds an option to speed up loading of pages that contain
multiple smaller images.
When the free memory is not enough to load the specified quality of the
picture, the automatic loading at a lower quality.
Move the Clean Invalid Bookmark function to the Options menu and display
the number of files involved in the invalid bookmarks being cleaned.
Improved support for JPEG2000 format images. In the Jasper library based
on the JPEG2000 decoder has been modified to use less memory but the
decoding speed has declined. You can use the original or modified
decoder in Options.
Add several common but previous versions do not contain the Chinese
code, you can open the use of these encoded PDF files.
Fixed the character overlap due to the incorrect width of the characters
in some fonts.
Fixed some problem with the position offset of some symbols.
Fixed some Chinese punctuation (such as '...') can not be displayed or
displayed as a box problem.
Support the use of bookr 0.71 original version of the CJK plug-in fonts,
in order to get better font display. If you want to use the plug-in